Extend deploy trigger to do the actual infra creation (#189)
