+++ modules/cloudfront-proxy-config/outputs.tf
@@ -9,3 +9,11 @@ output "table_arn" {
 output "table_name" {
   value = var.multiple_deployments ? aws_dynamodb_table.proxy_config[0].name : null
 }
+
+output "bucket_name" {
+  value = aws_s3_bucket.proxy_config_store.id
+}
+
+output "bucket_arn" {
+  value = aws_s3_bucket.proxy_config_store.arn
+}
