+++ modules/statics-deploy/variables.tf
@@ -2,6 +2,10 @@ variable "static_files_archive" {
   type = string
 }
 
+variable "static_files_archive_name" {
+  type = string
+}
+
 variable "deploy_trigger_module_version" {
   type    = string
   default = "0.4.0"
@@ -31,6 +35,12 @@ variable "use_awscli_for_static_upload" {
   default = false
 }
 
+variable "multiple_deployments" {
+  description = "Have multiple deployments and domain aliases."
+  type        = bool
+  default     = false
+}
+
 ###########
 # SQS Queue
 ###########
@@ -54,6 +64,60 @@ variable "deployment_name" {
 variable "tags" {
   type    = map(string)
   default = {}
+}
+
+#####################
+# Deployment creation
+#####################
+variable "lambda_attach_to_vpc" {
+  type        = bool
+  description = "Set to true if the Lambda functions should be attached to a VPC. Use this setting if VPC resources should be accessed by the Lambda functions. When setting this to true, use vpc_security_group_ids and vpc_subnet_ids to specify the VPC networking. Note that attaching to a VPC would introduce a delay on to cold starts"
+  default     = false
+}
+
+variable "vpc_subnet_ids" {
+  type        = list(string)
+  description = "The list of VPC subnet IDs to attach the Lambda functions. lambda_attach_to_vpc should be set to true for these to be applied."
+  default     = []
+}
+
+variable "vpc_security_group_ids" {
+  type        = list(string)
+  description = "The list of Security Group IDs to be used by the Lambda functions. lambda_attach_to_vpc should be set to true for these to be applied."
+  default     = []
+}
+
+variable "lambda_environment_variables" {
+  type        = map(string)
+  description = "Map that defines environment variables for the Lambda Functions in Next.js."
+  default     = {}
+}
+
+variable "proxy_config_table_name" {
+  description = "Name of the DynamoDB table to store proxy configurations."
+  type        = string
+  default     = "tf-next-proxy-config"
+}
+
+variable "proxy_config_table_arn" {
+  description = "ARN of the DynamoDB table to store proxy configurations."
+  type        = string
+}
+
+variable "proxy_config_bucket_name" {
+  description = "Name of the S3 bucket to store proxy configurations."
+  type        = string
+  default     = "next-tf-proxy-config"
+}
+
+variable "proxy_config_bucket_arn" {
+  description = "ARN of the S3 bucket to store proxy configurations."
+  type        = string
+}
+
+variable "lambda_logging_policy_arn" {
+  description = "ARN of the lambda logging policy."
+  type        = string
 }
 
 #######
